backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
dnspython>=2.4.0
pydantic>=2.5.0
